added stat client
