+++ main-service/src/main/java/ru/practicum/ewm/MainServer.java
@@ -1,0 +1,13 @@
+package ru.practicum.ewm;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MainServer {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MainServer.class, args);
+    }
+
+}
