+++ main-service/src/main/java/ru/practicum/ewm/category/controller/CategoryPublicController.java
@@ -1,0 +1,11 @@
+package ru.practicum.ewm.category.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/categories")
+@RequiredArgsConstructor
+public class CategoryPublicController {
+}
