+++ main-service/src/main/java/ru/practicum/ewm/category/mapper/CategoryMapper.java
@@ -1,0 +1,18 @@
+package ru.practicum.ewm.category.mapper;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.MappingConstants;
+import org.mapstruct.ReportingPolicy;
+import ru.practicum.ewm.category.Category;
+import ru.practicum.ewm.category.dto.CategoryDto;
+
+import java.util.List;
+
+@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = MappingConstants.ComponentModel.SPRING)
+public interface CategoryMapper {
+    Category toEntity(CategoryDto categoryDto);
+
+    CategoryDto toCategoryDto(Category category);
+
+    List<CategoryDto> toCategoryDtoList( List<Category> categoryList);
+}
