+++ main-service/src/main/java/ru/practicum/ewm/category/repository/CategoryRepository.java
@@ -1,0 +1,26 @@
+package ru.practicum.ewm.category.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+import ru.practicum.ewm.category.Category;
+
+import java.util.List;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+    boolean existsByNameIgnoreCaseAndIdNot(String name, Long id);
+
+    boolean existsByNameIgnoreCase(String name);
+
+    @Query(value = """
+            select *
+            from categories as c
+            order by c.id offset :from limit :size
+            """, nativeQuery = true)
+    List<Category> findCategories(@Param("from") int from,
+                                  @Param("size") int size);
+
+}
