+++ main-service/src/main/java/ru/practicum/ewm/exception/NotFoundException.java
@@ -1,0 +1,11 @@
+package ru.practicum.ewm.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+
+    public NotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
