+++ main-service/src/main/java/ru/practicum/ewm/exception/ValidationException.java
@@ -1,0 +1,7 @@
+package ru.practicum.ewm.exception;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(String message) {
+        super(message);
+    }
+}
