--- main-service/src/main/java/ru/practicum/ewm/main/MainServer.java
@@ -1,13 +1,0 @@
-package ru.practicum.ewm.main;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MainServer {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MainServer.class, args);
-    }
-
-}
