--- main-service/src/main/java/ru/practicum/ewm/main/category/controller/CategoryAdminController.java
@@ -1,37 +1,0 @@
-package ru.practicum.ewm.main.category.controller;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewm.main.category.dto.CategoryDto;
-import ru.practicum.ewm.main.category.service.CategoryService;
-
-@Slf4j
-@RestController
-@RequestMapping("/admin/categories")
-@RequiredArgsConstructor
-public class CategoryAdminController {
-
-    private final CategoryService categoryService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public CategoryDto createCategory(CategoryDto categoryDto) {
-        log.info("POST new category {}", categoryDto);
-        return categoryService.createCategory(categoryDto);
-    }
-
-    @PatchMapping("/{id}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public CategoryDto updateCategory(@PathVariable Long id, CategoryDto categoryDto) {
-        log.info("PATCH category id={}, {}", id, categoryDto);
-        return categoryService.updateCategory(id, categoryDto);
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCategory(@PathVariable Long id) {
-        categoryService.deleteCategory(id);
-    }
-}
