--- main-service/src/main/java/ru/practicum/ewm/main/category/controller/CategoryPublicController.java
@@ -1,11 +1,0 @@
-package ru.practicum.ewm.main.category.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/categories")
-@RequiredArgsConstructor
-public class CategoryPublicController {
-}
