--- main-service/src/main/java/ru/practicum/ewm/main/category/mapper/CategoryMapper.java
@@ -1,18 +1,0 @@
-package ru.practicum.ewm.main.category.mapper;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
-import org.mapstruct.ReportingPolicy;
-import ru.practicum.ewm.main.category.Category;
-import ru.practicum.ewm.main.category.dto.CategoryDto;
-
-import java.util.List;
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = MappingConstants.ComponentModel.SPRING)
-public interface CategoryMapper {
-    Category toEntity(CategoryDto categoryDto);
-
-    CategoryDto toCategoryDto(Category category);
-
-    List<CategoryDto> toCategoryDtoList( List<Category> categoryList);
-}
