--- main-service/src/main/java/ru/practicum/ewm/main/category/repository/CategoryRepository.java
@@ -1,14 +1,0 @@
-package ru.practicum.ewm.main.category.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.ewm.main.category.Category;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-
-    boolean existsByNameIgnoreCaseAndIdNot(String name, Long id);
-
-    boolean existsByNameIgnoreCase(String name);
-
-}
