--- main-service/src/main/java/ru/practicum/ewm/main/exception/NotFoundException.java
@@ -1,11 +1,0 @@
-package ru.practicum.ewm.main.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-
-    public NotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
