--- main-service/src/main/java/ru/practicum/ewm/main/exception/ValidationException.java
@@ -1,7 +1,0 @@
-package ru.practicum.ewm.main.exception;
-
-public class ValidationException extends RuntimeException {
-    public ValidationException(String message) {
-        super(message);
-    }
-}
