--- main-service/src/main/java/ru/practicum/ewm/main/user/controller/UserAdminController.java
@@ -1,46 +1,0 @@
-package ru.practicum.ewm.main.user.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewm.main.user.dto.UserDto;
-import ru.practicum.ewm.main.user.service.UserService;
-
-import java.util.List;
-
-@Slf4j
-@RestController
-@RequestMapping("/admin/users")
-@RequiredArgsConstructor
-public class UserAdminController {
-
-    private final UserService userService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDto createUser(@RequestBody @Valid UserDto userDto) {
-        log.info("POST new user {}", userDto);
-        return userService.createUser(userDto);
-    }
-
-    @GetMapping
-    public List<UserDto> getUsers(@RequestParam(name = "ids", defaultValue = "") List<Long> ids,
-                                  @RequestParam(name = "from", defaultValue = "0") int from,
-                                  @RequestParam(name = "size", defaultValue = "10") int size) {
-        log.info("GET users ids={}, from={}, size={}", ids, from, size);
-        if (ids == null || ids.isEmpty()) {
-            return userService.getUsers(from, size);
-        }
-        return userService.getUsersByIds(ids, from, size);
-    }
-
-    @DeleteMapping("/{userId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable(name = "userId") Long userId) {
-        log.info("DELETE user id={}", userId);
-        userService.deleteUserById(userId);
-    }
-
-}
