--- main-service/src/main/java/ru/practicum/ewm/main/user/mapper/UserMapper.java
@@ -1,21 +1,0 @@
-package ru.practicum.ewm.main.user.mapper;
-
-import org.mapstruct.*;
-import ru.practicum.ewm.main.user.User;
-import ru.practicum.ewm.main.user.dto.UserDto;
-import ru.practicum.ewm.main.user.dto.UserShortDto;
-
-import java.util.List;
-
-@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = MappingConstants.ComponentModel.SPRING)
-public interface UserMapper {
-    User toEntity(UserDto userDto);
-
-    UserDto toDto(User user);
-
-    List<UserDto> toDtoList(List<User> userList);
-
-    UserShortDto toShortDto(User user);
-
-    List<UserShortDto> toShortDtoList(List<User> userList);
-}
