--- main-service/src/main/java/ru/practicum/ewm/main/user/service/UserService.java
@@ -1,17 +1,0 @@
-package ru.practicum.ewm.main.user.service;
-
-import ru.practicum.ewm.main.user.dto.UserDto;
-
-import java.util.List;
-
-public interface UserService {
-
-    UserDto createUser(UserDto userDto);
-
-    List<UserDto> getUsersByIds(List<Long> ids, int from, int size);
-
-    List<UserDto> getUsers(int from, int size);
-
-    void deleteUserById(Long userId);
-
-}
