+++ main-service/src/main/resources/application.properties
@@ -1,4 +1,5 @@
 server.port=8080
+stats-server.url=http://localhost:9090
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
