+++ stats-service/stats-client/pom.xml
@@ -10,6 +10,7 @@
     </parent>
 
     <artifactId>stats-client</artifactId>
+    <packaging>jar</packaging>
 
     <properties>
         <maven.compiler.source>21</maven.compiler.source>
